push: 2022/08/11 周四 22:44

+++ P2011.cpp
@@ -1,14 +1,14 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-int a,b,k,n,m;
-int x[1010][1010];
+long long a,b,k,n,m;
+long long x[1010][1010];
 
 int main() {
   cin>>a>>b>>k>>n>>m;
   x[0][1] = 1;
-  for(int i = 1; i <= k; i++) {
-    for(int j = 1; j <= i + 1; j++) {
+  for(long long i = 1; i <= k; i++) {
+    for(long long j = 1; j <= i + 1; j++) {
       x[i][j] = (x[i-1][j-1]*b + x[i-1][j]*a)%10007;
     }
   }
